+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.6.2] - 2025-12-31
+
 ### Changed
 
 - Changed default connection linewidth from 2.5 to 1.5 for finer connection lines.
@@ -189,6 +191,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Virtual environment setup and dependency management
 - MIT License
 
+[0.6.2]: https://github.com/rob-mosher/musicxml-to-png/releases/tag/v0.6.2
 [0.6.1]: https://github.com/rob-mosher/musicxml-to-png/releases/tag/v0.6.1
 [0.6.0]: https://github.com/rob-mosher/musicxml-to-png/releases/tag/v0.6.0
 [0.5.0]: https://github.com/rob-mosher/musicxml-to-png/releases/tag/v0.5.0